FIX: 앱이 켜져있어도 fcm 푸시 받도록 수정
jira ref: #S12P21A204-150

+++ mobile/app/src/main/java/com/ssafy/jangan_mobile/service/dto/BeaconNotificationDto.kt
@@ -3,8 +3,8 @@ package com.ssafy.jangan_mobile.service.dto
 data class BeaconNotificationDto(
     val beaconName: String,
     val beaconCode: Int,
-    val coordX: Int,
-    val coordY: Int,
+    val coordX: Double,
+    val coordY: Double,
     val floor: Int,
     val imageUrl: String,
     val isNewFire: Int
